[Diagnostic] Allow multi-digit arguments in diagnostics.
This patch allows diagnostic arguments to be greater than 9.
Originally only 0-9 were allowed, the rational being any diagnostic
with more than 10 arguments needs to be simplifier. However, with
the introduction of substitutions it seems possible for this to no
longer be the case; A substitution may need 5 arguments, and the
diagnostic may need another 5.

The changes required in ClangDiagnosticEmitter.cpp are contained in
D46540.